webapp/requirements.txt
sentence-transformers
pinecone-client
transformers
pandas
torch